requirements.txt
requests
django
python-dotenv
djangorestframework
django-filter
